+++ zcash_local_net/src/validator/zebrad.rs
@@ -4,7 +4,7 @@ use crate::{
     config,
     error::LaunchError,
     launch,
-    logs::{self, LogsToDir, LogsToStdoutAndStderr as _},
+    logs::{LogsToDir, LogsToStdoutAndStderr as _},
     network,
     process::Process,
     utils::{
@@ -279,7 +279,6 @@ impl Process for Zebrad {
 
         let mut handle = command.spawn().expect(EXPECT_SPAWN);
 
-        logs::write_logs(&mut handle, &logs_dir);
         launch::wait(
         ProcessId::Zebrad,
         &mut handle,
@@ -469,14 +468,6 @@ impl Validator for Zebrad {
         }
         self.poll_chain_height(chain_height + n).await;
 
-        Ok(())
-    }
-
-    async fn generate_blocks_with_delay(&self, blocks: u32) -> std::io::Result<()> {
-        for _ in 0..blocks {
-            self.generate_blocks(1).await.unwrap();
-            tokio::time::sleep(std::time::Duration::from_millis(1500)).await;
-        }
         Ok(())
     }